+++ common/nodejs-utils/Cargo.toml
@@ -11,7 +11,8 @@ anyhow = "1"
 async-trait = "0.1.88"
 bon = "3"
 chrono = { version = "0.4", default-features = false, features = ["serde"] }
-bullet_stream = "0.8"
+bullet_stream = "0.10"
+futures = { version = "0.3", default-features = false, features = ["io-compat"] }
 indoc = "2"
 keep_a_changelog_file = "0.1.0"
 libcnb-data = "=0.29.0"
@@ -30,6 +31,7 @@ serde-xml-rs = "0.8"
 sha2 = "0.10.9"
 thiserror = "2"
 tokio = { version = "1", default-features = false, features = ["rt-multi-thread", "time", "macros"] }
+tokio-util = { version = "0.7", default-features = false, features = ["compat", "io"] }
 toml = "0.8"
 ureq = { version = "3", features = ["json"] }
 url = "2"
