+++ common/nodejs-utils/src/lib.rs
@@ -5,8 +5,8 @@ pub mod application;
 pub mod available_parallelism;
 pub mod buildplan;
 pub mod distribution;
-pub mod download_file;
 pub mod error_handling;
+pub mod http;
 pub mod inv;
 mod nodejs_org;
 mod npmjs_org;
